feat: remove Material Navigator, switch to localstorage for auth persistence.
