+++ examples/auth_flow/lib/main.dart
@@ -1,67 +1,64 @@
 import 'package:auth_flow/notifiers/auth_notifier.dart';
 import 'package:auth_flow/ui/login_page.dart';
 import 'package:auth_flow/ui/home_page.dart';
+import 'package:auth_flow/ui/profile_page.dart';
 import 'package:disco/disco.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_solidart/flutter_solidart.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:go_router/go_router.dart';
+import 'package:localstorage/localstorage.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  final prefs = await SharedPreferences.getInstance();
+  await initLocalStorage();
 
-  runApp(
-    ProviderScope(
-      providers: [sharedPreferenceProvider(prefs)],
-      child: ProviderScope(providers: [AuthNotifier.provider], child: const MyApp()),
-    ),
-  );
+  runApp(const MyApp());
 }
 
-class MyApp extends StatefulWidget {
+class MyApp extends StatelessWidget {
   const MyApp({super.key});
-
-  static final navigatorKey = GlobalKey<NavigatorState>();
-
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  late final Effect authEffect;
-
-  @override
-  void initState() {
-    super.initState();
-    // Listen to auth state changes and navigate accordingly
-    authEffect = Effect(() {
-      final isLoggedIn = AuthNotifier.provider.of(context).isLoggedIn.value;
-      if (!isLoggedIn) {
-        MyApp.navigatorKey.currentState?.popUntil((route) => route.isFirst);
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    authEffect();
-    super.dispose();
-  }
 
   @override
   Widget build(BuildContext context) {
-    return SignalBuilder(
-      builder: (context, child) {
-        return MaterialApp(
-          navigatorKey: MyApp.navigatorKey,
-          title: 'Auth Demo',
-          debugShowCheckedModeBanner: false,
-          theme: ThemeData(colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple)),
-          home: AuthNotifier.provider.of(context).isLoggedIn.value
-              ? const HomePage(title: 'Home')
-              : const LoginPage(),
-        );
-      },
+    return ProviderScope(
+      providers: [AuthNotifier.provider],
+      child: Builder(
+        builder: (context) {
+          return MaterialApp.router(
+            title: 'Auth Demo - GoRouter',
+            debugShowCheckedModeBanner: false,
+            theme: ThemeData(colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple)),
+            routerConfig: AppRouter(context).router,
+          );
+        },
+      ),
     );
   }
+}
+
+class AppRouter {
+  AppRouter(this.context);
+
+  final BuildContext context;
+
+  late final router = GoRouter(
+    refreshListenable: AuthNotifier.provider.of(context).isLoggedIn,
+    redirect: (context, state) {
+      final isLoggedIn = AuthNotifier.provider.of(context).isLoggedIn.value;
+      if (!isLoggedIn && state.matchedLocation != '/login') {
+        return '/login';
+      }
+      if (isLoggedIn && state.matchedLocation == '/login') {
+        return '/';
+      }
+      return null;
+    },
+    routes: [
+      GoRoute(
+        path: '/',
+        builder: (_, _) => const HomePage(title: 'Home'),
+        routes: [GoRoute(path: 'profile', builder: (_, _) => const ProfilePage())],
+      ),
+      GoRoute(path: '/login', builder: (_, _) => const LoginPage()),
+    ],
+  );
 }
