+++ examples/auth_flow/lib/ui/login_page.dart
@@ -1,3 +1,4 @@
+import 'package:auth_flow/domain/user.dart';
 import 'package:auth_flow/notifiers/auth_notifier.dart';
 import 'package:flutter/material.dart';
 
@@ -15,7 +16,7 @@ class LoginPage extends StatelessWidget {
             ElevatedButton(
               onPressed: () {
                 final controller = AuthNotifier.provider.of(context);
-                controller.login((id: '1', name: 'John Doe', email: 'john.doe@example.com'));
+                controller.login(User(id: '1', name: 'John Doe', email: 'john.doe@example.com'));
               },
               child: Text('Login'),
             ),
